+++ packages/react-app/src/components/HelperComponents/UserSubmissions.jsx
@@ -1,6 +1,7 @@
 import axios from "axios";
 import { lazy, Suspense, useEffect, useState } from "react";
 import { useHistory, useLocation } from "react-router-dom";
+import illustrationImage from "../../assets/illustration.png";
 
 const server = "https://talentdao-api.herokuapp.com";
 
@@ -9,7 +10,7 @@ const SubmissionCard = lazy(() => import("./SubmissionCard.jsx"));
 const UserSubmissions = ({ address }) => {
   // const [toArticles, goToArticles] = useState(false);
   const [articles, setArticles] = useState([]);
-  const [selectedOption, setSelectedOption] = useState();
+  
   const history = useHistory();
   const location = useLocation();
 
@@ -25,72 +26,62 @@ const UserSubmissions = ({ address }) => {
         }
       } catch (e) {
         console.log(e);
-      }
+      } 
     };
     getArticles();
   }, [address]);
 
-  const options = [{ name: "interest" }, { name: "interest" }];
-  const onSelectChange = e => {
-    console.log(e.target.value);
-    setSelectedOption(e.target.value);
-  };
-
   return (
     <div className="flex flex-col bg-white p-8">
       {articles.length === 0 ? (
-        <div className="flex flex-row mb-6 place-content-between">
-          <div className="flex flex-row">
-            <svg width="17" height="12" viewBox="0 0 17 12" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <path
-                d="M0.469669 5.46967C0.176777 5.76256 0.176777 6.23744 0.469669 6.53033L5.24264 11.3033C5.53553 11.5962 6.01041 11.5962 6.3033 11.3033C6.59619 11.0104 6.59619 10.5355 6.3033 10.2426L2.06066 6L6.3033 1.75736C6.59619 1.46447 6.59619 0.989593 6.3033 0.696699C6.01041 0.403806 5.53553 0.403806 5.24264 0.696699L0.469669 5.46967ZM17 5.25L1 5.25V6.75L17 6.75V5.25Z"
-                fill="black"
-              />
-            </svg>
-            <div className="ml-1 -mt-0.5 font-bold cursor-pointer">Back</div>
-          </div>
-          <div className="flex flex-row space-x-4">
-            <div className="flex flex-row">
-              <select
-                className="block bg-transparent w-full pl-3 pr-3 py-1 text-lg rounded-2xl"
-                style={{ border: "1px solid #E6E6E6" }}
-                onChange={e => onSelectChange(e)}
+        <div className="flex flex-col mb-6 place-content-between mt-4">
+          <div className="flex flex-row mb-6 place-content-between">
+            <div className="ml-1 -mt-1 font-bold cursor-pointer text-xl">Submissions</div>
+            <div className="flex flex-row space-x-4">
+              <div
+                className="ml-1 -mt-0.5 cursor-pointer text-lg text-primary"
+                onClick={() => history.push(`/user/publisher`)}
               >
-                {options.map((item, index) => {
-                  return <option>{item.name}</option>;
-                })}
-              </select>
-            </div>
-            <div className="flex flex-row space-x-1 rounded-full p-px bg-grey">
-              <div className="px-6 rounded-full text-lg bg-white border border-grey cursor-pointer flex flex-col items-center">
-                Reviewer
+                See Reviewers & Publishers
               </div>
-              <div className="px-6 pt-1 rounded-full text-lg text-darkgray cursor-pointer flex flex-col items-center">
-                Publisher
-              </div>
+              <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+                <path
+                  d="M4 12H20M13 5L20 12L13 19"
+                  stroke="#B41C2E"
+                  stroke-width="2"
+                  stroke-linecap="round"
+                  stroke-linejoin="round"
+                />
+              </svg>
             </div>
           </div>
-          <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-3 gap-4">
-            {articles.map((item, index) => {
+          <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-4">
+            {/* {articles.map((item, index) => {
               return (
                 <Suspense fallback={<div>Loading Submission Card...</div>}>
                   <SubmissionCard key={index} article={item}></SubmissionCard>
                 </Suspense>
               );
-            })}
+            })} */}
+            <SubmissionCard></SubmissionCard>
+            <SubmissionCard></SubmissionCard>
+            <SubmissionCard></SubmissionCard>
+            <SubmissionCard></SubmissionCard>
+            <SubmissionCard></SubmissionCard>
+            <SubmissionCard></SubmissionCard>
+            <SubmissionCard></SubmissionCard>
           </div>
         </div>
       ) : (
-        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-3 gap-4">
-          <div className="flex justify-center">
-            <div className="rounded-2xl p-8 bg-white flex flex-col space-y-4">
-              <div
-                className="rounded-xl text-lg bg-primary text-white text-center cursor-pointer px-4 py-2"
-                onClick={() => history.push(`/submit/${address}`)}
-              >
-                Make A Submission
-              </div>
-            </div>
+        <div className="flex flex-col justify-center py-52">
+          <img className="self-center w-1/5 mb-6" src={illustrationImage} />
+          <div className="text-center text-xl font-bold mb-1">Nothing to see here</div>
+          <div className="text-center text-base mb-6">Upload your next article, document on Talent DAO</div>
+          <div
+            className="w-1/5 self-center rounded-full text-lg bg-primary text-white text-center cursor-pointer px-4 py-4"
+            onClick={() => history.push(`/submit/${address}`)}
+          >
+            Make A Submission
           </div>
         </div>
       )}
